+++ plugins/destination/bigquery/cloud-config-ui/src/components/guide/index.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+import { AuthType, useFormContext } from '@cloudquery/plugin-config-ui-lib';
+
+import { IAMGuide } from './iam-guide';
+import { ServiceAccountGuide } from './sa-guide';
+
+export function Guide() {
+  const form = useFormContext();
+  const authType = form.watch('_authType');
+
+  return authType === AuthType.OAUTH ? <IAMGuide /> : <ServiceAccountGuide />;
+}