+++ plugins/source/hackernews/cloud-config-ui/src/utils/prepareSubmitValues.ts
@@ -1,11 +1,16 @@
 import { PluginUiMessagePayload } from '@cloudquery/plugin-config-ui-connector';
-import { corePrepareSubmitValues, PluginTable } from '@cloudquery/plugin-config-ui-lib';
+import {
+  corePrepareSubmitValues,
+  PluginConfig,
+  PluginTable,
+} from '@cloudquery/plugin-config-ui-lib';
 
 export function prepareSubmitValues(
+  config: PluginConfig,
   values: Record<string, any>,
   tablesList?: PluginTable[],
 ): PluginUiMessagePayload['validation_passed']['values'] {
-  const payload = corePrepareSubmitValues(values, tablesList);
+  const payload = corePrepareSubmitValues(config, values, tablesList);
 
   if (values.item_concurrency) {
     payload.spec.item_concurrency = Number(values.item_concurrency);