+++ scripts/update-codegen.sh
@@ -1,6 +1,6 @@
 PLUGIN_DIRS=$(ls plugins/source)
 for plugin in $PLUGIN_DIRS; do
-	if [ ! -d "plugins/source/$plugin/codegen" ]; then
+	if ! grep -q "gen:" "plugins/source/$plugin/Makefile"; then
 	  continue;
 	fi
 	echo "Generating code for $plugin"