+++ plugins/destination/azblob/CHANGELOG.md
@@ -1,5 +1,22 @@
 # Changelog
 
+## [4.2.5](https://github.com/cloudquery/cloudquery/compare/plugins-destination-azblob-v4.2.4...plugins-destination-azblob-v4.2.5) (2024-10-16)
+
+
+### Bug Fixes
+
+* **deps:** Update module github.com/Azure/azure-sdk-for-go/sdk/azidentity to v1.8.0 ([#19366](https://github.com/cloudquery/cloudquery/issues/19366)) ([16c8e99](https://github.com/cloudquery/cloudquery/commit/16c8e99d5c4b6fe5ff564171698e49edc3f31165))
+* **deps:** Update module github.com/cloudquery/plugin-sdk/v4 to v4.66.1 ([#19380](https://github.com/cloudquery/cloudquery/issues/19380)) ([0b37067](https://github.com/cloudquery/cloudquery/commit/0b3706722e10da4e5f065b86927c555df5fd350f))
+
+## [4.2.4](https://github.com/cloudquery/cloudquery/compare/plugins-destination-azblob-v4.2.3...plugins-destination-azblob-v4.2.4) (2024-10-08)
+
+
+### Bug Fixes
+
+* **deps:** Update module github.com/cloudquery/plugin-sdk/v4 to v4.64.0 ([#19287](https://github.com/cloudquery/cloudquery/issues/19287)) ([49941ee](https://github.com/cloudquery/cloudquery/commit/49941ee0c985fe6cb88581818064c8152a388304))
+* **deps:** Update module github.com/cloudquery/plugin-sdk/v4 to v4.65.0 ([#19322](https://github.com/cloudquery/cloudquery/issues/19322)) ([87a68ea](https://github.com/cloudquery/cloudquery/commit/87a68ea489068a621948112137f987252b83273c))
+* **deps:** Update module github.com/cloudquery/plugin-sdk/v4 to v4.66.0 ([#19332](https://github.com/cloudquery/cloudquery/issues/19332)) ([137a232](https://github.com/cloudquery/cloudquery/commit/137a2328637ef226e5dba446b92c2d670f798540))
+
 ## [4.2.3](https://github.com/cloudquery/cloudquery/compare/plugins-destination-azblob-v4.2.2...plugins-destination-azblob-v4.2.3) (2024-09-25)