+++ plugins/source/xkcd/CHANGELOG.md
@@ -1,5 +1,30 @@
 # Changelog
 
+## [1.4.3](https://github.com/cloudquery/cloudquery/compare/plugins-source-xkcd-v1.4.2...plugins-source-xkcd-v1.4.3) (2024-10-16)
+
+
+### Bug Fixes
+
+* **deps:** Update module github.com/cloudquery/plugin-sdk/v4 to v4.66.1 ([#19380](https://github.com/cloudquery/cloudquery/issues/19380)) ([0b37067](https://github.com/cloudquery/cloudquery/commit/0b3706722e10da4e5f065b86927c555df5fd350f))
+
+## [1.4.2](https://github.com/cloudquery/cloudquery/compare/plugins-source-xkcd-v1.4.1...plugins-source-xkcd-v1.4.2) (2024-10-09)
+
+
+### Bug Fixes
+
+* **deps:** Update module github.com/cloudquery/plugin-sdk/v4 to v4.64.0 ([#19287](https://github.com/cloudquery/cloudquery/issues/19287)) ([49941ee](https://github.com/cloudquery/cloudquery/commit/49941ee0c985fe6cb88581818064c8152a388304))
+* **deps:** Update module github.com/cloudquery/plugin-sdk/v4 to v4.65.0 ([#19322](https://github.com/cloudquery/cloudquery/issues/19322)) ([87a68ea](https://github.com/cloudquery/cloudquery/commit/87a68ea489068a621948112137f987252b83273c))
+* **deps:** Update module github.com/cloudquery/plugin-sdk/v4 to v4.66.0 ([#19332](https://github.com/cloudquery/cloudquery/issues/19332)) ([137a232](https://github.com/cloudquery/cloudquery/commit/137a2328637ef226e5dba446b92c2d670f798540))
+* **deps:** Update module google.golang.org/grpc to v1.67.1 ([#19328](https://github.com/cloudquery/cloudquery/issues/19328)) ([ce09a61](https://github.com/cloudquery/cloudquery/commit/ce09a61eb208bdf9022ca8fb966ff6356fd8d688))
+
+## [1.4.1](https://github.com/cloudquery/cloudquery/compare/plugins-source-xkcd-v1.4.0...plugins-source-xkcd-v1.4.1) (2024-10-02)
+
+
+### Bug Fixes
+
+* **deps:** Update dependency @types/react to v18.3.10 ([#19260](https://github.com/cloudquery/cloudquery/issues/19260)) ([22cc644](https://github.com/cloudquery/cloudquery/commit/22cc644be1951973c5ad3637a286fb08961cc4a7))
+* **deps:** Update dependency eslint to v8.57.1 ([#19264](https://github.com/cloudquery/cloudquery/issues/19264)) ([2baa628](https://github.com/cloudquery/cloudquery/commit/2baa62862a4015528a5748949728bac21854d9b5))
+
 ## [1.4.0](https://github.com/cloudquery/cloudquery/compare/plugins-source-xkcd-v1.3.7...plugins-source-xkcd-v1.4.0) (2024-09-25)